Kernel hyperparameter tuning, grid operation
Changes to hyperparameter (kernel) selection function tuneKernel to avoid local minima with small datasets.
Full support for grid-search BO.
Updated gentype to include nominally constant variables. So for example if you set the lengthscale in the kernel to 1.2c (note the c for nominally constant) then tuneKernel will leave it alone and not try to tune it.