Python integration
Added a python frontend - pyheavy. To compile install pybind11, set the required directories in the makefile and make pybasic or make pyoptmath.

Also tweaked parsing code for vectors and sets to allow for non-spaced input (eg [1 2]) and the inclusion of commas (eg [1,2] is now valid syntax for a vector, in line with python syntax).

Plus the usual minor bug fixes.